updated daily command
